+++ app/controllers/works_controller.rb
@@ -248,8 +248,8 @@ class WorksController < ApplicationController # rubocop:disable Metrics/ClassLen
     return true if @version_status.open? && (deposit? || initial_review?)
 
     # If no previous change, then check for a current change.
-    mapped_cocina_object = ToCocina::Work::Mapper.call(work_form: @work_form, content: @content,
-                                                       source_id: "h3:object-#{@work.id}")
+    mapped_cocina_object = Cocina::WorkMapper.call(work_form: @work_form, content: @content,
+                                                   source_id: "h3:object-#{@work.id}")
     RoundtripSupport.changed?(cocina_object: mapped_cocina_object)
   end
