+++ app/roundtrippers/collection_roundtripper.rb
@@ -32,8 +32,8 @@ class CollectionRoundtripper
   attr_reader :collection_form, :content
 
   def roundtripped_cocina_object
-    ToCocina::Collection::Mapper.call(collection_form:,
-                                      source_id: normalized_original_cocina_object.identification&.sourceId)
+    Cocina::CollectionMapper.call(collection_form:,
+                                  source_id: normalized_original_cocina_object.identification&.sourceId)
   end
 
   def normalized_original_cocina_object
