+++ app/roundtrippers/work_roundtripper.rb
@@ -34,9 +34,9 @@ class WorkRoundtripper
   attr_reader :work_form, :content
 
   def roundtripped_cocina_object
-    ToCocina::Work::Mapper.call(work_form:,
-                                content:,
-                                source_id: normalized_original_cocina_object.identification&.sourceId)
+    Cocina::WorkMapper.call(work_form:,
+                            content:,
+                            source_id: normalized_original_cocina_object.identification&.sourceId)
   end
 
   def normalized_original_cocina_object
